requirements.txt
# Telegram Bot Framework
aiogram==3.4.1
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.13.1  # Database migrations
pgvector==0.2.4  # Vector search extension

# AI APIs
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.18.1

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
telethon==1.34.0  # For Telegram scraping

# Scheduling
apscheduler==3.10.4

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Configuration
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Logging
loguru==0.7.2

# Testing (optional)
pytest==8.0.0
pytest-asyncio==0.23.4
